#176 refactor: import元のファイルをstore/index.tsにする（同様の内容をexportしているので、こっちのほうがわかりやすい）
